requirements.txt
pyrogram
tgcrypto
google-genai
python-dotenv
aiohttp
yt-dlp
yandex-music
selenium
pillow
beautifulsoup4
lxml
openpyxl
psutil
telegraph
markdown
requests
qrcode
fastapi
uvicorn
jinja2
aiofiles